refactor: move callback to workspace manager
